Cleanup and optimize code
Clean up test projects

Removed dead code and cleaned up a bit

Convert affix command lookup to an enum type

Cleanup code

+++ src/WeCantSpell.Hunspell/AffixConfigOptions.cs
@@ -7,7 +7,7 @@ namespace WeCantSpell.Hunspell
     /// </summary>
     /// <seealso cref="AffixConfig"/>
     [Flags]
-    public enum AffixConfigOptions : int
+    public enum AffixConfigOptions : short
     {
         /// <summary>
         /// Indicates that no options are set.
